requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
joblib==1.3.2
pillow==10.2.0